Add files via upload
AtliQ Grands is a company that owns multiple five-star hotels across India. They have been in the hospitality industry for the past 20 years. Due to strategic moves from other competitors and ineffective decision-making, the management noticed that they are losing their market share and revenue. As a strategic move, the managing director of AtliQ Grands decided to incorporate “Business and Data Intelligence” to regain their market share and revenue.